Some more small cleanups

+++ src/hash.hpp
@@ -13,7 +13,7 @@ struct djb2_hash {
     std::size_t calc(std::size_t hash, const char *str) const noexcept {
         std::size_t c;
 
-        while ((c = *str++)) {
+        while ((c = static_cast<std::size_t>(*str++))) {
             hash = ((hash << 5U) + hash) + c; /* hash * 33 + c */
         }
